requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-gsheets==0.0.10
djangorestframework==3.15.1
djangorestframework-firebase==0.3.0
drf-yasg==1.21.7
environ==1.0
firebase-admin==2.18.0
gcloud==0.18.3
google-api-core==1.34.1
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.48.2
gsheets==0.6.1
gspread==6.1.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.6
msgpack==1.0.8
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pyparsing==3.1.2
Pyrebase4==4.7.1
python-decouple==3.8
python-dotenv==1.0.1
python-jwt==4.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.29.0
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9
serpy==0.3.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
StrEnum==0.4.15
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18
whitenoise==6.6.0
